Add a test for user-agent parsing
